FastQC data extraction
I added the code to extract extra info from FastQC. I also changed some variable names to separate the info from the summary file and the data file, which are two different outputs from FastQC. This still needs to be tested to make sure it works. Please take special attention to the json, I am not familiar with the structure so it is possible I didn't get it completely right.